Refactored namespaces. Added better error handling for missing config files.
