+++ src/Auth/Model/Application.php
@@ -3,12 +3,12 @@
  * Copyright 2016 Andrew O'Rourke
  */
 
-namespace FUM8\Auth\Model;
+namespace Auth\Model;
 
 /**
- * @package FUM8\Auth\Model
+ * @package Auth\Model
  */
-class Character extends Base
+class Application extends Base
 {
     public static $_table = 'applications';
 }
