+++ src/Auth/Model/AuthToken.php
@@ -3,10 +3,10 @@
  * Copyright 2016 Andrew O'Rourke
  */
 
-namespace FUM8\Auth\Model;
+namespace Auth\Model;
 
 /**
- * @package FUM8\Auth\Model
+ * @package Auth\Model
  */
 class AuthToken extends Base
 {
