+++ src/Auth/Model/Character.php
@@ -3,12 +3,12 @@
  * Copyright 2016 Andrew O'Rourke
  */
 
-namespace FUM8\Auth\Model;
+namespace Auth\Model;
 
 /**
- * @package FUM8\Auth\Model
+ * @package Auth\Model
  */
-class Application extends Base
+class Character extends Base
 {
     public static $_table = 'applications';
 }
