+++ views/front/index.html.php
@@ -12,7 +12,9 @@
     <script src="<?=getVersionedAsset('js/jquery-1.12.3.min.js');?>"></script>
     <script src="<?=getVersionedAsset('js/bootstrap.min.js');?>"></script>
     <script src="<?=getVersionedAsset('js/login.js');?>"></script>
+    <?php if (isset($_GET['redirect'])) : ?>
     <script type="text/javascript">window.redirectURI = "<?=base64_decode($_GET['redirect'])?>";</script>
+    <?php endif ?>
 </head>
 <body>
     <div id="login-box-container">
